Extract Matrix3 and Vector3 classes
Let's take a moment to talk about JavaScript/TypeScript bullshit. In
JS/TS, you can access object properties by doing this:

    obj["foo"];

Which is apparently how array access is done:

    arr[0];

What this means is that if you refactor an object, such that what was
once accessed like this:

    obj.arr[0];

is now accessed like this:

    obj.subobj.arr[0];

Nothing will help you with that refactor. Nothing. You're fucked. You
have to find each and every instance of `obj.arr[0]` and replace it by
hand with `obj.subobj.arr[0]` because JS is fucking spoonfed its own
breakfast and TS is too fucking cowardly to try and actually fix
meaningful failures in the language it's trying to replace.

Fuck all this bullshit.